+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.numbergenerator;
+package racingcar.domain;
 
 import java.util.Random;
