+++ src/test/java/racingcar/domain/car/CarPositionTest.java
@@ -2,6 +2,7 @@ package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.entity.vo.CarPosition;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
