+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -3,6 +3,7 @@ package racingcar.domain.car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.entity.car.Car;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
